+++ CHANGELOG.md
@@ -79,6 +79,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Switch to auto-generated `String` and `Equal` methods for most proto messages [#1957](https://github.com/provenance-io/provenance/pull/1957).
 * Clean up the marker module's expected BankKeeper interface [#1954](https://github.com/provenance-io/provenance/pull/1954).
 
+### Deprecated
+
+* The `provenanced query account` command has been removed. It is still available as `provenanced query auth account` TODO: Add link.
+
 ### Dependencies
 
 - Bump `github.com/cosmos/ibc-go/v8` from 8.0.0 to 8.2.1 ([#1910](https://github.com/provenance-io/provenance/pull/1910), [#1956](https://github.com/provenance-io/provenance/pull/1956))